backend/requirements.txt
# Versiones probadas en Ubuntu 24.04 / Python 3.12
# Puedes actualizar libremente.

Flask==3.0.3
Flask-JWT-Extended==4.6.0
Flask-Mail==0.9.1
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
marshmallow==3.21.2
marshmallow-sqlalchemy==1.0.0
itsdangerous==2.2.0
python-dotenv==1.0.1

# Base de datos
psycopg2-binary==2.9.9

# Tareas asíncronas
celery==5.4.0
redis==5.0.4

# Linter (opcional, desarrollo)
ruff==0.4.4
